Actualizando a Postgresql

+++ database/migrations/2018_03_06_024616_create_proveedores_table.php
@@ -14,12 +14,12 @@ class CreateProveedoresTable extends Migration
     public function up()
     {
         Schema::create('proveedores', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigIncrements('id')->unsigned();
             $table->string('contacto',50)->nullable();
-            $table->string('telefono_contacto',50)->nullable(); 
-            
-            $table->foreign('id')->references('id')->on('personas')->onDelete('cascade'); 
-            
+            $table->string('telefono_contacto',50)->nullable();
+
+            $table->foreign('id')->references('id')->on('personas')->onDelete('cascade');
+
         });
     }
