+++ database/migrations/2018_03_14_000000_create_users_table.php
@@ -14,9 +14,9 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigIncrements('id')->unsigned();
             $table->foreign('id')->references('id')->on('personas')->onDelete('cascade');
-            
+
             $table->string('usuario')->unique();
             $table->string('password');
             $table->boolean('condicion')->default(1);
